Clean up code for emoji picker
