+++ src/components/Posts/ComposePost.tsx
@@ -1,13 +1,12 @@
 import { Box, Button, IconButton, Stack, TextField } from "@mui/material";
 import AddPhotoAlternateOutlinedIcon from "@mui/icons-material/AddPhotoAlternateOutlined";
-import EmojiEmotionsOutlinedIcon from "@mui/icons-material/EmojiEmotionsOutlined";
-import { useState, useEffect, useRef } from "react";
+import { useState } from "react";
 import axios from "axios";
 import { useAppDispatch, useAppSelector } from "../../state/hooks";
 import { appendPost } from "../../state/slices/postsSlice";
 import UserAvatar from "../Common/UserAvatar";
-import EmojiPicker, { EmojiStyle, EmojiClickData } from "emoji-picker-react";
-import ClickOffEmojis from "../Common/ClickOffEmojiPicker";
+import { EmojiClickData } from "emoji-picker-react";
+import EmojiPickerIconButton from "../Common/EmojiPickerIconButton";
 
 type ComposePostProps = {
   placeholder: string;
@@ -30,17 +29,10 @@ const styles = {
     justifyContent: "space-between",
     marginTop: 1,
   },
-  emojiContainer: { position: "absolute", zIndex: 1 },
-  emojiPicker: {
-    height: 300,
-    width: 250,
-    "--epr-emoji-size": "25px",
-  },
 };
 
 const ComposePost = ({ placeholder, minRows }: ComposePostProps) => {
   const [postTextContent, setPostTextContent] = useState("");
-  const [showEmojiPicker, setShowEmojiPicker] = useState(false);
   const user = useAppSelector((state) => state.user);
   const dispatch = useAppDispatch();
 
@@ -65,10 +57,6 @@ const ComposePost = ({ placeholder, minRows }: ComposePostProps) => {
     }
   };
 
-  // useEffect(() => {
-  //   console.log(showEmojiPicker);
-  // }, [showEmojiPicker]);
-
   return (
     <form onSubmit={onSubmit}>
       <Box sx={styles.compostPostContainer}>
@@ -92,21 +80,13 @@ const ComposePost = ({ placeholder, minRows }: ComposePostProps) => {
               <IconButton size="small">
                 <AddPhotoAlternateOutlinedIcon />
               </IconButton>
-              <Stack direction="column">
-                <IconButton
-                  id="emoji-button"
-                  size="small"
-                  onClick={(event) => {
-                    event.stopPropagation();
-                    if (!showEmojiPicker) {
-                      console.log("test");
-                      setShowEmojiPicker(true);
-                    }
-                  }}
-                >
-                  <EmojiEmotionsOutlinedIcon />
-                </IconButton>
-              </Stack>
+              <EmojiPickerIconButton
+                onEmojiClick={(emoji: EmojiClickData) => {
+                  setPostTextContent(
+                    (prevContent) => prevContent + emoji.emoji
+                  );
+                }}
+              />
             </Stack>
             <Button
               disabled={!postTextContent.trim()}
@@ -117,16 +97,6 @@ const ComposePost = ({ placeholder, minRows }: ComposePostProps) => {
               Post
             </Button>
           </Box>
-
-          <ClickOffEmojis
-            setPostContent={(emoji: EmojiClickData) => {
-              setPostTextContent((prevContent) => prevContent + emoji.emoji);
-            }}
-            setShowEmojiPicker={() => {
-              setShowEmojiPicker(false);
-            }}
-            emojiContainerStyle={styles.emojiContainer}
-          />
         </Box>
       </Box>
     </form>
